#FDF-15 Merging Lean Memory changes to 3.0 branch TESTS_RAN: Unit test, dev test Corrected some warnings in merged checkin
git-svn-id: https://10.196.60.217/svn/schooner-products/branches/fdf-3.0@2434 01a69087-22d5-4a29-8152-8a1d5e10e5e9

+++ api/tests/FDF_Async_Deletion2.c
@@ -9,7 +9,7 @@
 static char *base = "container";
 static int iterations = 200000;
 static int threads = 2;
-static long size = 1024 * 1024 * 1024 * 4;
+static long size = (uint64_t) 1024 * 1024 * 1024 * 4;
 
 pthread_mutex_t count_mutex   = PTHREAD_MUTEX_INITIALIZER;
 pthread_cond_t  condition_var = PTHREAD_COND_INITIALIZER;
